Fix sum bug in kd computation.

+++ fairseq/optim/lr_scheduler/cosine_lr_schedule.py
@@ -82,5 +82,5 @@ class CosineSchedule(FairseqLRScheduler):
     def cosine_cycle(self, num_updates):
         if num_updates <= self.args.warmup_updates:
             return -1
-        return (num_updates - self.args.warmup_updates) / self.args.cosine_cycle_steps
+        return (int)((num_updates - self.args.warmup_updates) / self.args.cosine_cycle_steps)
